Runtime: Move signal handler to main and init it only once
Before, the signal handler was initiated _every_ time a new set of
robot code and carburator children was spawned. However, the handler handles
any signal (Alrm in our case) that is sent to the pid of runtime itself.
So, because we spawned more than one sighandler thread and had no way of killing them,
they would accumulate. One of them would handle the message, but there was no way of
knowing which one.

This change resolves this issue. The message is sent over a channel from the sighandler
and this message is received in the tcp read loop right before the call to read_exact.
This should prevent most deadlocks.