Runtime: Allow multiple Cockpit-Backend connections
Previously, when you reconnected with runtime from Cockpit-Backend,
Runtime would close as only one connection was being listened on.
When that connection was closed, the program would exit.

I'm using the crossbeam crate to be able to use a MPMC model.
This way we can always have a single ALRM signal listener in the program.
If you have a better suggestion that doen not use another crate your
welcome to give feedback, but I'm not sure how to propperly do it the standard
Rust way.